% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptor-RedisContainer.R
\name{getServerContainer,RedisContainer-method}
\alias{getServerContainer,RedisContainer-method}
\title{Get the Redis server container}
\usage{
\S4method{getServerContainer}{RedisContainer}(workerContainer)
}
\arguments{
\item{workerContainer}{The worker container.}
}
\value{
A \code{RedisContainer} server container
}
\description{
Get the Redis server container from the worker container
}
