% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA.R
\docType{class}
\name{RedisContainer-class}
\alias{RedisContainer-class}
\alias{.RedisContainer}
\title{The base redis container}
\description{
The base redis container. It should not be directly used by the user.
}

