% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-RedisContainer.R
\name{RedisContainer-commom-parameters}
\alias{RedisContainer-commom-parameters}
\title{Common RedisContainer parameter}
\arguments{
\item{image}{Character, the container image}

\item{name}{Character, the optional name of the container}

\item{environment}{List, the environment variables in the container}

\item{tag}{Character, the image tag}
}
\value{
No reuturn value
}
\description{
Common RedisContainer parameter
}
